FIXED: Fix warning

+++ Sources/libMultiMarkdown/epub.c
@@ -255,7 +255,7 @@ void epub_export_nav_entry(DString * out, const char * source, scratch_pad * scr
 
 		if (entry_level >= level) {
 			// This entry is a direct descendant of the parent
-			scratch->label_counter = *counter;
+			scratch->label_counter = (int)*counter;
 			temp_char = label_from_header(source, entry, scratch);
 			printf("<li><a href=\"main.xhtml#%s\">", temp_char);
 			mmd_export_token_tree_html(out, source, entry->child, scratch);
